+++ src/deluge/gui/views/session_view.cpp
@@ -22,7 +22,9 @@
 #include "gui/colour/colour.h"
 #include "gui/colour/palette.h"
 #include "gui/context_menu/audio_input_selector.h"
-#include "gui/context_menu/launch_style.h"
+#include "gui/context_menu/clip_settings/launch_style.h"
+#include "gui/context_menu/clip_settings/new_clip_type.h"
+#include "gui/context_menu/context_menu.h"
 #include "gui/context_menu/stem_export/cancel_stem_export.h"
 #include "gui/menu_item/colour.h"
 #include "gui/ui/keyboard/keyboard_screen.h"
@@ -72,6 +74,7 @@
 #include "storage/audio/audio_file_manager.h"
 #include "storage/file_item.h"
 #include "storage/storage_manager.h"
+#include "task_scheduler.h"
 #include "util/cfunctions.h"
 #include "util/d_string.h"
 #include "util/functions.h"
@@ -90,6 +93,7 @@ SessionView sessionView{};
 
 SessionView::SessionView() {
 	xScrollBeforeFollowingAutoExtendingLinearRecording = -1;
+	newClipToCreate = nullptr;
 }
 
 bool SessionView::getGreyoutColsAndRows(uint32_t* cols, uint32_t* rows) {
@@ -469,27 +473,19 @@ moveAfterClipInstance:
 				}
 			}
 			else if (currentUIMode == UI_MODE_HOLDING_STATUS_PAD) {
-				context_menu::launchStyle.setupAndCheckAvailability();
-				openUI(&context_menu::launchStyle);
+				context_menu::clip_settings::launchStyle.setupAndCheckAvailability();
+				openUI(&context_menu::clip_settings::launchStyle);
 			}
 			else if (currentUIMode == UI_MODE_CLIP_PRESSED_IN_SONG_VIEW) {
 				actionLogger.deleteAllLogs();
 				performActionOnPadRelease = false;
 
 				Clip* clip = getClipForLayout();
-				if (currentSong->sessionLayout == SessionLayoutType::SessionLayoutTypeGrid) {
-					requestRendering(this, 0xFFFFFFFF, 0xFFFFFFFF);
-					if (gridModeActive == SessionGridModeConfig) {
-						context_menu::launchStyle.clip = clip;
-						context_menu::launchStyle.setupAndCheckAvailability();
-						openUI(&context_menu::launchStyle);
-					}
-					else if (clip != nullptr) {
-						replaceInstrumentClipWithAudioClip(clip);
-					}
-				}
-				else if (clip != nullptr) {
-					replaceInstrumentClipWithAudioClip(clip);
+				requestRendering(this, 0xFFFFFFFF, 0xFFFFFFFF);
+				if (clip != nullptr) {
+					context_menu::clip_settings::launchStyle.clip = clip;
+					context_menu::clip_settings::launchStyle.setupAndCheckAvailability();
+					openUI(&context_menu::clip_settings::launchStyle);
 				}
 			}
 			else if (currentUIMode == UI_MODE_NONE) {
@@ -794,7 +790,15 @@ startHoldingDown:
 					// if (possiblyCreatePendingNextOverdub(clipIndex, OverdubType::EXTENDING)) return
 					// ActionResult::DEALT_WITH;
 
-					clip = createNewInstrumentClip(yDisplay);
+					context_menu::clip_settings::newClipType.yDisplay = yDisplay;
+					context_menu::clip_settings::newClipType.setupAndCheckAvailability();
+					openUI(&context_menu::clip_settings::newClipType);
+
+					// wait until you've exited out of clip creation context menu
+					newClipToCreate = nullptr;
+					yield([]() { return (getCurrentUI() != &context_menu::clip_settings::newClipType); });
+
+					clip = newClipToCreate;
 					if (!clip) {
 						return ActionResult::DEALT_WITH;
 					}
@@ -811,7 +815,6 @@ startHoldingDown:
 					clipWasSelectedWithShift = Buttons::isShiftButtonPressed();
 					selectedClipYDisplay = clipIndex - currentSong->songViewYScroll;
 					requestRendering(this, 0, 1 << selectedClipYDisplay);
-					goto startHoldingDown;
 				}
 			}
 
@@ -1577,50 +1580,100 @@ Error setPresetOrNextUnlaunchedOne(InstrumentClip* clip, OutputType outputType, 
 constexpr float colourStep = 22.5882352941;
 static float lastColour = 192 - colourStep + 1;
 
-Clip* SessionView::createNewInstrumentClip(int32_t yDisplay) {
+Clip* SessionView::createNewClip(OutputType outputType, int32_t yDisplay) {
+	Clip* clip = nullptr;
+	switch (outputType) {
+	case OutputType::AUDIO:
+		if (currentSong->sessionLayout == SessionLayoutType::SessionLayoutTypeGrid) {
+			clip = gridCreateAudioClipWithNewTrack();
+		}
+		else {
+			clip = createNewAudioClip(yDisplay);
+		}
+		break;
+	default:
+		if (currentSong->sessionLayout == SessionLayoutType::SessionLayoutTypeGrid) {
+			clip = gridCreateInstrumentClipWithNewTrack(outputType);
+		}
+		else {
+			clip = createNewInstrumentClip(outputType, yDisplay);
+		}
+		break;
+	}
+	return clip;
+}
+
+Clip* SessionView::createNewAudioClip(int32_t yDisplay) {
 	actionLogger.deleteAllLogs();
 
-	void* memory = GeneralMemoryAllocator::get().allocMaxSpeed(sizeof(InstrumentClip));
-	if (memory == nullptr) {
+	// Allocate memory for audio clip
+	void* clipMemory = GeneralMemoryAllocator::get().allocMaxSpeed(sizeof(AudioClip));
+	if (clipMemory == nullptr) {
 		display->displayError(Error::INSUFFICIENT_RAM);
 		return nullptr;
 	}
 
-	InstrumentClip* newClip = new (memory) InstrumentClip(currentSong);
+	// Create the audio clip and ParamManager
+	AudioClip* newClip = new (clipMemory) AudioClip();
 
-	uint32_t currentDisplayLength = currentSong->xZoom[NAVIGATION_CLIP] * kDisplayWidth;
-
-	if (playbackHandler.playbackState
-	    && (currentPlaybackMode == &arrangement || !playbackHandler.isEitherClockActive())) {
-		newClip->activeIfNoSolo = false;
+	// suss output
+	if (!createNewTrackForAudioClip(newClip)) {
+		newClip->~AudioClip();
+		delugeDealloc(clipMemory);
+		display->displayError(Error::INSUFFICIENT_RAM);
+		return nullptr;
 	}
 
-	uint32_t oneBar = currentSong->getBarLength();
+	// Give the new clip its stuff
+	setupNewClip(newClip);
 
-	// Default Clip length. Default to current zoom, minimum 1 bar
-	int32_t newClipLength = std::max(currentDisplayLength, oneBar);
+	// Insert and Resync New Clip
+	if (!insertAndResyncNewClip(newClip, yDisplay)) {
+		newClip->~AudioClip();
+		delugeDealloc(clipMemory);
+		display->displayError(Error::INSUFFICIENT_RAM);
+		return nullptr;
+	}
 
-	newClip->colourOffset = random(72);
-	newClip->loopLength = newClipLength;
+	return newClip;
+}
 
-	bool instrumentAlreadyInSong;
+Clip* SessionView::createNewInstrumentClip(OutputType outputType, int32_t yDisplay) {
+	actionLogger.deleteAllLogs();
 
-	OutputType outputType = OutputType::SYNTH;
-doGetInstrument:
-	Error error = setPresetOrNextUnlaunchedOne(newClip, outputType, &instrumentAlreadyInSong);
-	if (error != Error::NONE) {
+	// Allocate memory for instrument clip
+	void* clipMemory = GeneralMemoryAllocator::get().allocMaxSpeed(sizeof(InstrumentClip));
+	if (clipMemory == nullptr) {
+		display->displayError(Error::INSUFFICIENT_RAM);
+		return nullptr;
+	}
 
-		// If that was for a synth and there were none, try a kit
-		if (error == Error::NO_FURTHER_PRESETS && outputType == OutputType::SYNTH) {
-			outputType = OutputType::KIT;
-			goto doGetInstrument;
-		}
+	// create the instrument clip and param manager
+	InstrumentClip* newClip = new (clipMemory) InstrumentClip(currentSong);
+
+	// suss output
+	if (!createNewTrackForInstrumentClip(outputType, newClip, true)) {
 		newClip->~InstrumentClip();
-		delugeDealloc(memory);
-		display->displayError(error);
-		return NULL;
+		delugeDealloc(clipMemory);
+		return nullptr;
 	}
 
+	// Give the new clip its stuff
+	setupNewClip(newClip);
+
+	// Insert and Resync New Clip
+	if (!insertAndResyncNewClip(newClip, yDisplay)) {
+		newClip->~InstrumentClip();
+		delugeDealloc(clipMemory);
+		display->displayError(Error::INSUFFICIENT_RAM);
+		return nullptr;
+	}
+
+	return newClip;
+}
+
+bool SessionView::insertAndResyncNewClip(Clip* newClip, int32_t yDisplay) {
+	// insert clip at index
 	int32_t index = yDisplay + currentSong->songViewYScroll;
 	if (index <= 0) {
 		index = 0;
@@ -1632,63 +1685,25 @@ doGetInstrument:
 		newClip->section =
 		    currentSong->sessionClips.getClipAtIndex(currentSong->sessionClips.getNumElements() - 1)->section;
 	}
-	currentSong->sessionClips.insertClipAtIndex(newClip, index);
+	if (currentSong->sessionClips.insertClipAtIndex(newClip, index) != Error::NONE) {
+		return false;
+	}
 
+	// resync new clip play pos
 	char modelStackMemory[MODEL_STACK_MAX_SIZE];
 	ModelStack* modelStack = setupModelStackWithSong(modelStackMemory, currentSong);
 	ModelStackWithTimelineCounter* modelStackWithTimelineCounter = modelStack->addTimelineCounter(newClip);
 
+	resyncNewClip(newClip, modelStackWithTimelineCounter);
+
+	return true;
+}
+
+void SessionView::resyncNewClip(Clip* newClip, ModelStackWithTimelineCounter* modelStackWithTimelineCounter) {
 	// Figure out the play pos for the new Clip if we're currently playing
 	if (session.hasPlaybackActive() && playbackHandler.isEitherClockActive() && currentSong->isClipActive(newClip)) {
 		session.reSyncClip(modelStackWithTimelineCounter, true);
 	}
-
-	if (!instrumentAlreadyInSong) {
-		currentSong->addOutput(newClip->output);
-	}
-
-	// Possibly want to set this as the active Clip...
-	if (!newClip->output->getActiveClip()) {
-		newClip->output->setActiveClip(modelStackWithTimelineCounter);
-	}
-
-	return newClip;
-}
-
-void SessionView::replaceInstrumentClipWithAudioClip(Clip* clip) {
-	int32_t clipIndex = currentSong->sessionClips.getIndexForClip(clip);
-
-	if (!clip || clip->type != ClipType::INSTRUMENT) {
-		return;
-	}
-
-	if (currentSong->sessionLayout == SessionLayoutType::SessionLayoutTypeGrid
-	    && currentSong->getClipWithOutput(clip->output, false, clip)) {
-		display->displayPopup(deluge::l10n::get(
-		    deluge::l10n::String::STRING_FOR_INSTRUMENTS_WITH_CLIPS_CANT_BE_TURNED_INTO_AUDIO_TRACKS));
-		return;
-	}
-
-	// don't allow clip type change if clip is not empty
-	InstrumentClip* instrumentClip = (InstrumentClip*)clip;
-	if (!instrumentClip->isEmpty()) {
-		return;
-	}
-
-	Clip* newClip = currentSong->replaceInstrumentClipWithAudioClip(clip, clipIndex);
-
-	if (!newClip) {
-		display->displayError(Error::INSUFFICIENT_RAM);
-		return;
-	}
-
-	currentSong->arrangementYScroll--; // Is our best bet to avoid the scroll appearing to change visually
-
-	view.setActiveModControllableTimelineCounter(newClip);
-	view.displayOutputName(newClip->output, true, newClip);
-
-	// If Clip was in keyboard view, need to redraw that
-	requestRendering(this, 1 << selectedClipYDisplay, 1 << selectedClipYDisplay);
 }
 
 void SessionView::removeClip(Clip* clip) {
@@ -2909,11 +2924,6 @@ void SessionView::gridRenderActionModes(int32_t y, RGB image[][kDisplayWidth + k
 		modeColour = colours::blue; // Blue
 		break;
 	}
-	case GridMode::YELLOW: {
-		modeActive = (gridModeActive == SessionGridModeConfig);
-		modeColour = colours::yellow; // Yellow
-		break;
-	}
 	case GridMode::PINK: {
 		modeActive = performanceSessionView.gridModeActive;
 		modeColour = colours::magenta; // Pink
@@ -3030,13 +3040,7 @@ RGB SessionView::gridRenderClipColor(Clip* clip) {
 		clip->output->colour = lastColour;
 	}
 
-	RGB resultColour;
-	if (gridModeActive == SessionGridModeConfig) {
-		resultColour = view.getClipMuteSquareColour(clip, resultColour, true, false);
-	}
-	else {
-		resultColour = RGB::fromHue(clip->output->colour);
-	}
+	RGB resultColour = RGB::fromHue(clip->output->colour);
 
 	// If we are not in record arming mode make this clip full color for being soloed
 	if ((clip->soloingInSessionMode || clip->armState == ArmState::ON_TO_SOLO) && !viewingRecordArmingActive) {
@@ -3101,19 +3105,26 @@ Clip* SessionView::gridCreateClipInTrack(Output* targetOutput) {
 	actionLogger.deleteAllLogs();
 
 	// For safety we set it up exactly as we want it
-	newClip->colourOffset = random(72);
-	newClip->loopLength = currentSong->getBarLength();
-	newClip->activeIfNoSolo = false;
-	newClip->soloingInSessionMode = false;
-	newClip->wasActiveBefore = false;
-	newClip->isPendingOverdub = false;
-	newClip->isUnfinishedAutoOverdub = false;
-	newClip->armState = ArmState::OFF;
+	setupNewClip(newClip);
 
 	return newClip;
 }
 
-bool SessionView::gridCreateNewTrackForClip(OutputType type, InstrumentClip* clip, bool copyDrumsFromClip) {
+bool SessionView::createNewTrackForAudioClip(AudioClip* newClip) {
+	// Suss output
+	AudioOutput* newOutput = currentSong->createNewAudioOutput();
+	if (!newOutput) {
+		return false;
+	}
+
+	char modelStackMemory[MODEL_STACK_MAX_SIZE];
+	ModelStack* modelStack = setupModelStackWithSong(modelStackMemory, currentSong);
+	newClip->setOutput(modelStack->addTimelineCounter(newClip), newOutput);
+
+	return true;
+}
+
+bool SessionView::createNewTrackForInstrumentClip(OutputType type, InstrumentClip* clip, bool copyDrumsFromClip) {
 	bool instrumentAlreadyInSong = false;
 	if (type == OutputType::SYNTH || type == OutputType::KIT) {
 		Error error = setPresetOrNextUnlaunchedOne(clip, type, &instrumentAlreadyInSong, copyDrumsFromClip);
@@ -3157,7 +3168,29 @@ bool SessionView::gridCreateNewTrackForClip(OutputType type, InstrumentClip* cli
 	return true;
 }
 
-InstrumentClip* SessionView::gridCreateClipWithNewTrack(OutputType type) {
+AudioClip* SessionView::gridCreateAudioClipWithNewTrack() {
+	// Allocate new clip
+	void* memory = GeneralMemoryAllocator::get().allocMaxSpeed(sizeof(AudioClip));
+	if (!memory) {
+		display->displayError(Error::INSUFFICIENT_RAM);
+		return nullptr;
+	}
+
+	AudioClip* newClip = new (memory) AudioClip();
+	if (!createNewTrackForAudioClip(newClip)) {
+		newClip->~AudioClip();
+		delugeDealloc(memory);
+		display->displayError(Error::INSUFFICIENT_RAM);
+		return nullptr;
+	}
+
+	// For safety we set it up exactly as we want it
+	setupNewClip(newClip);
+
+	return newClip;
+}
+
+InstrumentClip* SessionView::gridCreateInstrumentClipWithNewTrack(OutputType type) {
 	// Allocate new clip
 	void* memory = GeneralMemoryAllocator::get().allocMaxSpeed(sizeof(InstrumentClip));
 	if (!memory) {
@@ -3166,23 +3199,61 @@ InstrumentClip* SessionView::gridCreateClipWithNewTrack(OutputType type) {
 	}
 
 	InstrumentClip* newClip = new (memory) InstrumentClip(currentSong);
-	if (!gridCreateNewTrackForClip(type, newClip, true)) {
+	if (!createNewTrackForInstrumentClip(type, newClip, true)) {
 		newClip->~InstrumentClip();
 		delugeDealloc(memory);
 		return nullptr;
 	}
 
 	// For safety we set it up exactly as we want it
+	setupNewClip(newClip);
+
+	return newClip;
+}
+
+// For safety we set it up the new clip exactly as we want it
+void SessionView::setupNewClip(Clip* newClip) {
 	newClip->colourOffset = random(72);
-	newClip->loopLength = currentSong->getBarLength();
-	newClip->activeIfNoSolo = false;
 	newClip->soloingInSessionMode = false;
 	newClip->wasActiveBefore = false;
 	newClip->isPendingOverdub = false;
 	newClip->isUnfinishedAutoOverdub = false;
 	newClip->armState = ArmState::OFF;
 
-	return newClip;
+	if (currentSong->sessionLayout == SessionLayoutType::SessionLayoutTypeGrid) {
+		newClip->loopLength = currentSong->getBarLength();
+		newClip->activeIfNoSolo = false;
+	}
+	else {
+		uint32_t currentDisplayLength = currentSong->xZoom[NAVIGATION_CLIP] * kDisplayWidth;
+		uint32_t oneBar = currentSong->getBarLength();
+
+		// Default Clip length. Default to current zoom, minimum 1 bar
+		int32_t newClipLength = std::max(currentDisplayLength, oneBar);
+
+		newClip->loopLength = newClipLength;
+
+		if (playbackHandler.playbackState
+		    && (currentPlaybackMode == &arrangement || !playbackHandler.isEitherClockActive())) {
+			newClip->activeIfNoSolo = false;
+		}
+	}
+
+	if (newClip->type == ClipType::AUDIO) {
+		if (currentSong->defaultAudioClipOverdubOutputCloning == -1) {
+			currentSong->defaultAudioClipOverdubOutputCloning = 1;
+			// For each Clip in session
+			for (int32_t c = 0; c < currentSong->sessionClips.getNumElements(); c++) {
+				Clip* clip = currentSong->sessionClips.getClipAtIndex(c);
+
+				if (clip->type == ClipType::AUDIO && clip->armedForRecording) {
+					currentSong->defaultAudioClipOverdubOutputCloning = ((AudioClip*)clip)->overdubsShouldCloneOutput;
+					break;
+				}
+			}
+		}
+		newClip->overdubsShouldCloneOutput = currentSong->defaultAudioClipOverdubOutputCloning;
+	}
 }
 
 Clip* SessionView::gridCreateClip(uint32_t targetSection, Output* targetOutput, Clip* sourceClip) {
@@ -3217,7 +3288,13 @@ Clip* SessionView::gridCreateClip(uint32_t targetSection, Output* targetOutput, 
 	// Create new clip in new track
 	else {
 		// This is the right position to add immediate type creation
-		newClip = gridCreateClipWithNewTrack(OutputType::SYNTH);
+		context_menu::clip_settings::newClipType.setupAndCheckAvailability();
+		openUI(&context_menu::clip_settings::newClipType);
+
+		// wait until you've exited out of clip creation context menu
+		newClipToCreate = nullptr;
+		yield([]() { return (getCurrentUI() != &context_menu::clip_settings::newClipType); });
+		newClip = newClipToCreate;
 	}
 
 	// Set new clip section and add it to the list
@@ -3248,7 +3325,7 @@ Clip* SessionView::gridCreateClip(uint32_t targetSection, Output* targetOutput, 
 			InstrumentClip* newInstrumentClip = (InstrumentClip*)newClip;
 			// Create a new track for the clip
 			if (targetOutput == nullptr) {
-				if (!gridCreateNewTrackForClip(sourceClip->output->type, newInstrumentClip, false)) {
+				if (!createNewTrackForInstrumentClip(sourceClip->output->type, newInstrumentClip, false)) {
 					currentSong->sessionClips.deleteAtIndex(0);
 					newClip->~Clip();
 					delugeDealloc(newClip);
@@ -3292,9 +3369,7 @@ Clip* SessionView::gridCreateClip(uint32_t targetSection, Output* targetOutput, 
 	}
 
 	// Figure out the play pos for the new Clip if we're currently playing
-	if (session.hasPlaybackActive() && playbackHandler.isEitherClockActive() && currentSong->isClipActive(newClip)) {
-		session.reSyncClip(modelStack, true);
-	}
+	resyncNewClip(newClip, modelStack);
 
 	// Set to active for new tracks
 	if (targetOutput == nullptr && !newClip->output->getActiveClip()) {
@@ -3380,10 +3455,6 @@ ActionResult SessionView::gridHandlePads(int32_t x, int32_t y, int32_t on) {
 				gridModeActive = SessionGridModeEdit;
 				break;
 			}
-			case GridMode::YELLOW: {
-				gridModeActive = SessionGridModeConfig;
-				break;
-			}
 			case GridMode::PINK: {
 				performanceSessionView.gridModeActive = true;
 				performanceSessionView.timeGridModePress = AudioEngine::audioSampleTimer;
@@ -3418,10 +3489,6 @@ ActionResult SessionView::gridHandlePads(int32_t x, int32_t y, int32_t on) {
 		}
 		case SessionGridModeLaunch: {
 			modeHandleResult = gridHandlePadsLaunch(x, y, on, clip);
-			break;
-		}
-		case SessionGridModeConfig: {
-			modeHandleResult = gridHandlePadsConfig(x, y, on, clip);
 			break;
 		}
 		}
@@ -3491,6 +3558,8 @@ ActionResult SessionView::gridHandlePadsEdit(int32_t x, int32_t y, int32_t on, C
 		return ActionResult::ACTIONED_AND_CAUSED_CHANGE;
 	}
 
+	bool createdNewClip = false;
+
 	if (on) {
 		// Only do this if no pad is pressed yet
 		if (gridFirstPressedX == -1 && gridFirstPressedY == -1) {
@@ -3509,6 +3578,7 @@ ActionResult SessionView::gridHandlePadsEdit(int32_t x, int32_t y, int32_t on, C
 					// Immediately start playing it for new tracks
 					if (clip != nullptr && track == nullptr) {
 						gridToggleClipPlay(clip, true);
+						createdNewClip = true;
 					}
 				}
 			}
@@ -3518,6 +3588,11 @@ ActionResult SessionView::gridHandlePadsEdit(int32_t x, int32_t y, int32_t on, C
 			}
 			// we've either created or selected a clip, so set it to be current
 			currentSong->setCurrentClip(clip);
+
+			// if we just created a new clip, treat it like it's not being pressed anymore
+			if (createdNewClip) {
+				return ActionResult::ACTIONED_AND_CAUSED_CHANGE;
+			}
 
 			// Allow clip control (selection)
 			currentUIMode = UI_MODE_CLIP_PRESSED_IN_SONG_VIEW;
@@ -3748,29 +3823,6 @@ void SessionView::gridHandlePadsLaunchToggleArming(Clip* clip, bool immediate) {
 			gridToggleClipPlay(clip, false);
 		}
 	}
-}
-
-ActionResult SessionView::gridHandlePadsConfig(int32_t x, int32_t y, int32_t on, Clip* clip) {
-	if (x < kDisplayWidth) {
-		if (on) {
-			if (gridFirstPressedX == -1 && gridFirstPressedY == -1) {
-				gridFirstPressedX = x;
-				gridFirstPressedY = y;
-			}
-
-			if (clip != nullptr) {
-				currentSong->setCurrentClip(clip);
-				currentUIMode = UI_MODE_CLIP_PRESSED_IN_SONG_VIEW;
-				view.displayOutputName(clip->output, true, clip);
-			}
-		}
-		else {
-			if (gridFirstPressedX == x && gridFirstPressedY == y) {
-				clipPressEnded();
-			}
-		}
-	}
-	return ActionResult::ACTIONED_AND_CAUSED_CHANGE;
 }
 
 ActionResult SessionView::gridHandleScroll(int32_t offsetX, int32_t offsetY) {